COMMIT
Update script.py

Was buying when stock was stable, which meant that when it had gone down, but it had been a couple of seconds, it was buying stock. It would then lose a couple of cents, and sell again, wait for it to be stable, and buy it again, to lose more cents and sell again. 

Code was changed to only buy when on the up and hold when stable at that point, as well as sell when going down. 

I lost a whole percent because of this error! Glad I fixed it. Everything works now.